Major Refactor
Major refactor was made to make it flexible and usable across other projects.
