+++ .gitignore
@@ -1,4 +1,12 @@
-
+# Default ignored files
+/shelf/
+/workspace.xml
+# Editor-based HTTP Client requests
+/httpRequests/
+# Datasource local storage ignored files
+/dataSources/
+/dataSources.local.xml
+# Obsidian
 .obsidian/app.json
 .obsidian/appearance.json
 .obsidian/core-plugins-migration.json
@@ -14,5 +22,6 @@
 .obsidian/core-plugins-migration.json
 .obsidian/core-plugins.json
 .obsidian/workspace.json
+# Others
 .env
-__init__.py
+hephaestus/__init__.py
