--- core/model/elysium/resources/media.py
@@ -1,1 +1,0 @@
-#News, articles references
