--- core/model/elysium/resources/title.py
@@ -1,1 +1,0 @@
-
