+++ examples/pipeline_00/config/sources.json
@@ -1,0 +1,6 @@
+[
+  {
+  },
+  {
+  }
+]
