+++ hephaestus/core/utils/general_helper.py
