--- pipelines/arg_census/config/config.yaml
@@ -1,6 +1,0 @@
-job_name: "ETL Job - Geopolitical Entities"
-
-app_code: "SLPdkGkoVEw"
-app_name: "hephaestus"
-
-pop_pipeline_code: "9QNCPw9v2Ki"
