--- pipelines/arg_census/config/sources.json
@@ -1,15 +1,0 @@
-[
-  {
-    "source_name": "2022 Argentinian Census",
-    "location_type": "FILE",
-    "location": "https://redatam.indec.gob.ar/redarg/tempo/96530/~tmp_9653001.xlsX",
-    "location_endpoint": null,
-    "extension": "xlsX",
-    "extract_type": "GET",
-    "params": null,
-    "headers": null,
-    "timeout": 50,
-    "pipeline_code": "9QNCPw9v2Ki"
-  }
-]
-
