--- pipelines/arg_parliament/config/config.yaml
@@ -1,8 +1,0 @@
-job_name: "ETL Job - Argentinian Parliament"
-
-app_code: "SLPdkGkoVEw"
-app_name: "hephaestus"
-
-upper_pipeline_code: "jYhFXgQk2vG"
-lower_pipeline_code: "DK95dHeya13"
-legislation_pipeline_code: "Rga1vGhfM4C"
