--- pipelines/arg_parliament/config/sources.json
@@ -1,218 +1,0 @@
-[
-  {
-    "source_name": "Lower Chamber Members",
-    "location_type": "API",
-    "location": "https://datos.hcdn.gob.ar/api/3/",
-    "location_endpoint": "action/datastore_search",
-    "extension": "",
-    "extract_type": "GET",
-    "params": {"resource_id": "169de2eb-465f-4007-a4c2-39a5ba4c0df3", "limit": 100, "offset": 0},
-    "headers": null,
-    "timeout": 50,
-    "pipeline_code": "DK95dHeya13"
-  },
-  {
-    "source_name": "Lower Chamber Committees",
-    "location_type": "API",
-    "location": "https://datos.hcdn.gob.ar/api/3/",
-    "location_endpoint": "action/datastore_search",
-    "extension": "",
-    "extract_type": "GET",
-    "params": {"resource_id": "6774a719-b439-4ef9-8943-9ea66bb588c6", "limit": 100, "offset": 0},
-    "headers": null,
-    "timeout": 50,
-    "pipeline_code": "DK95dHeya13"
-  },
-  {
-    "source_name": "Lower Chamber Committees Members",
-    "location_type": "API",
-    "location": "https://datos.hcdn.gob.ar/api/3/",
-    "location_endpoint": "action/datastore_search",
-    "extension": "",
-    "extract_type": "GET",
-    "params": {"resource_id": "8f360114-2663-43e3-99b2-e588560e5bec", "limit": 100, "offset": 0},
-    "headers": null,
-    "timeout": 50,
-    "pipeline_code": "DK95dHeya13"
-  },
-  {
-    "source_name": "Lower Chamber Committees Members Groups",
-    "location_type": "API",
-    "location": "https://datos.hcdn.gob.ar/api/3/",
-    "location_endpoint": "action/datastore_search",
-    "extension": "",
-    "extract_type": "GET",
-    "params": {"resource_id": "68bd5ebb-991f-4a4a-8ffd-4eee0fd608bb", "limit": 100, "offset": 0},
-    "headers": null,
-    "timeout": 50,
-    "pipeline_code": "DK95dHeya13"
-  },
-  {
-    "source_name": "Lower Chamber Referral Legislation",
-    "location_type": "API",
-    "location": "https://datos.hcdn.gob.ar/api/3/",
-    "location_endpoint": "action/datastore_search",
-    "extension": "",
-    "extract_type": "GET",
-    "params": {"resource_id": "b6774819-f8bc-44c7-b685-7da2e7fb3e55", "limit": 100, "offset": 0},
-    "headers": null,
-    "timeout": 50,
-    "pipeline_code": "Rga1vGhfM4C"
-  },
-  {
-    "source_name": "Lower Chamber Report Legislation",
-    "location_type": "API",
-    "location": "https://datos.hcdn.gob.ar/api/3/",
-    "location_endpoint": "action/datastore_search",
-    "extension": "",
-    "extract_type": "GET",
-    "params": {"resource_id": "59595a93-5a5e-4ba6-a3db-c1044e2f949e", "limit": 100, "offset": 0},
-    "headers": null,
-    "timeout": 50,
-    "pipeline_code": "Rga1vGhfM4C"
-  },
-  {
-    "source_name": "Lower Chamber Enacted Legislation",
-    "location_type": "API",
-    "location": "https://datos.hcdn.gob.ar/api/3/",
-    "location_endpoint": "action/datastore_search",
-    "extension": "",
-    "extract_type": "GET",
-    "params": {"resource_id": "68dfd7f8-91f3-4ecf-aebf-a860d1ca1a98", "limit": 100, "offset": 0},
-    "headers": null,
-    "timeout": 50,
-    "pipeline_code": "Rga1vGhfM4C"
-  },
-  {
-    "source_name": "Lower Chamber Legislation Projects",
-    "location_type": "API",
-    "location": "https://datos.hcdn.gob.ar/api/3/",
-    "location_endpoint": "action/datastore_search",
-    "extension": "",
-    "extract_type": "GET",
-    "params": {"resource_id": "22b2d52c-7a0e-426b-ac0a-a3326c388ba6", "limit": 100, "offset": 0},
-    "headers": null,
-    "timeout": 50,
-    "pipeline_code": "Rga1vGhfM4C"
-  },
-  {
-    "source_name": "Lower Chamber Legislation Projects Movement",
-    "location_type": "API",
-    "location": "https://datos.hcdn.gob.ar/api/3/",
-    "location_endpoint": "action/datastore_search",
-    "extension": "",
-    "extract_type": "GET",
-    "params": {"resource_id": "6108ea83-3f12-423c-a136-df1ae9cb2972", "limit": 100, "offset": 0},
-    "headers": null,
-    "timeout": 50,
-    "pipeline_code": "Rga1vGhfM4C"
-  },
-  {
-    "source_name": "Lower Chamber Projects Results",
-    "location_type": "API",
-    "location": "https://datos.hcdn.gob.ar/api/3/",
-    "location_endpoint": "action/datastore_search",
-    "extension": "",
-    "extract_type": "GET",
-    "params": {"resource_id": "daf0e90e-d94f-4186-b7cf-dfad5b6f9369", "limit": 100, "offset": 0},
-    "headers": null,
-    "timeout": 50,
-    "pipeline_code": "Rga1vGhfM4C"
-  },
-  {
-    "source_name": "Legislative Periods",
-    "location_type": "FILE",
-    "location": "https://datos.hcdn.gob.ar/dataset/b2081b7c-0fd5-4ea5-804d-daec6c97176e/resource/c342159d-263a-459c-bc53-2945afe33f74/download/periodos_paralamentarios1.5.json",
-    "location_endpoint": null,
-    "extension": "json",
-    "extract_type": "GET",
-    "params": null,
-    "headers": null,
-    "timeout": null,
-    "pipeline_code": "DK95dHeya13"
-  },
-  {
-    "source_name": "Lower Chamber Sessions",
-    "location_type": "FILE",
-    "location": "https://datos.hcdn.gob.ar/dataset/f744ea10-83b4-4493-8bef-6fc9fb9e41e9/resource/ed8783bc-a755-4227-ad07-fd8ace0aeb55/download/sesiones1.9.json",
-    "location_endpoint": null,
-    "extension": "json",
-    "extract_type": "GET",
-    "params": null,
-    "headers": null,
-    "timeout": null,
-    "pipeline_code": "DK95dHeya13"
-  },
-  {
-    "source_name": "Lower Chamber Publications",
-    "location_type": "FILE",
-    "location": "https://datos.hcdn.gob.ar/dataset/0467709c-f955-4218-a1ee-47c4d789f15a/resource/a775de5f-45ae-4eca-b35f-94fd5d8bef66/download/publicaciones1.11.json",
-    "location_endpoint": null,
-    "extension": "json",
-    "extract_type": "GET",
-    "params": null,
-    "headers": null,
-    "timeout": null,
-    "pipeline_code": "Rga1vGhfM4C"
-  },
-  {
-    "source_name": "Lower Chamber Staff",
-    "location_type": "FILE",
-    "location": "https://datos.hcdn.gob.ar/dataset/af7136d5-84eb-49b2-af2e-aaf6319e4aae/resource/3c819c65-35c2-4671-a4f6-4e557e677c92/download/nomina3.1.json",
-    "location_endpoint": null,
-    "extension": "json",
-    "extract_type": "GET",
-    "params": null,
-    "headers": null,
-    "timeout": null,
-    "pipeline_code": "DK95dHeya13"
-  },
-  {
-    "source_name": "Upper Chamber Members Historical",
-    "location_type": "FILE",
-    "location": "https://www.senado.gob.ar/micrositios/DatosAbiertos/ExportarListadoSenadoresHistorico/json",
-    "location_endpoint": null,
-    "extension": "json",
-    "extract_type": "GET",
-    "params": null,
-    "headers": null,
-    "timeout": null,
-    "pipeline_code": "jYhFXgQk2vG"
-  },
-  {
-    "source_name": "Upper Chamber Members",
-    "location_type": "FILE",
-    "location": "https://www.senado.gob.ar/micrositios/DatosAbiertos/ExportarListadoSenadores/json",
-    "location_endpoint": null,
-    "extension": "json",
-    "extract_type": "GET",
-    "params": null,
-    "headers": null,
-    "timeout": null,
-    "pipeline_code": "jYhFXgQk2vG"
-  },
-  {
-    "source_name": "Upper Chamber Staff",
-    "location_type": "FILE",
-    "location": "https://www.senado.gob.ar/micrositios/DatosAbiertos/ExportarListadoAgentes/json",
-    "location_endpoint": null,
-    "extension": "json",
-    "extract_type": "GET",
-    "params": null,
-    "headers": null,
-    "timeout": null,
-    "pipeline_code": "jYhFXgQk2vG"
-  },
-  {
-    "source_name": "Upper Chamber Contract Based Staff",
-    "location_type": "FILE",
-    "location": "https://www.senado.gob.ar/micrositios/DatosAbiertos/ExportarListadoContratos/json",
-    "location_endpoint": null,
-    "extension": "json",
-    "extract_type": "GET",
-    "params": null,
-    "headers": null,
-    "timeout": null,
-    "pipeline_code": "jYhFXgQk2vG"
-  }
-]
