--- pipelines/geopolitical_entities/config/config.yaml
@@ -1,7 +1,0 @@
-job_name: "ETL Job - Geopolitical Entities"
-
-app_code: "SLPdkGkoVEw"
-app_name: "hephaestus"
-
-countries_pipeline_code: "i7a9wvnzqH8"
-subdivisions_pipeline_code: "dcKPjEPQhfF"
