--- pipelines/geopolitical_entities/config/sources.json
@@ -1,39 +1,0 @@
-[
-  {
-    "source_name": "ISO 3166-1",
-    "location_type": "API",
-    "location": "https://public.opendatasoft.com/api/explore/v2.1/",
-    "location_endpoint": "catalog/datasets/countries-codes/records?",
-    "extension": null,
-    "extract_type": "GET",
-    "params": {"limit": 20, "offset": 0},
-    "headers": null,
-    "timeout": 50,
-    "pipeline_code": "i7a9wvnzqH8"
-  },
-  {
-    "source_name": "ISO 3166-2 Wiki Template",
-    "location_type": "WEB",
-    "location": "https://en.wikipedia.org/wiki/ISO_3166-2:{iso3166a2_code}",
-    "location_endpoint": null,
-    "extension": null,
-    "extract_type": "GET",
-    "params": null,
-    "headers": null,
-    "timeout": 50,
-    "pipeline_code": "dcKPjEPQhfF",
-    "active": false
-  },
-  {
-    "source_name": "ISO 3166-2 ARG",
-    "location_type": "FILE",
-    "location": "https://infra.datos.gob.ar/georef/provincias.json",
-    "location_endpoint": null,
-    "extension": "json",
-    "extract_type": "GET",
-    "params": null,
-    "headers": null,
-    "timeout": 50,
-    "pipeline_code": "dcKPjEPQhfF"
-  }
-]
