--- pipelines/ilostat/config/config.yaml
@@ -1,6 +1,0 @@
-job_name: "ETL Job - ILOSTAT"
-
-app_code: "SLPdkGkoVEw"
-app_name: "hephaestus"
-
-ISCO_pipeline_code: "UY3zhDM2P2X"
