--- pipelines/ilostat/config/sources.json
@@ -1,14 +1,0 @@
-[
-  {
-    "source_name": "ISCO-08",
-    "location_type": "FILE",
-    "location": "https://www.ilo.org/ilostat-files/ISCO/newdocs-08-2021/ISCO-08/ISCO-08%20EN%20Structure%20and%20definitions.xlsx",
-    "location_endpoint": null,
-    "extension": "xlsx",
-    "extract_type": "GET",
-    "params": null,
-    "headers": null,
-    "timeout": 50,
-    "pipeline_code": "UY3zhDM2P2X"
-  }
-]
